moved person element to inherit from base class -removed error prone indexing
